Refactor update tags
